#21: move entities to new package
