+++ src/main/java/com/devcomanda/easymeetup/repository/MeetupRepository.java
@@ -1,6 +1,6 @@
 package com.devcomanda.easymeetup.repository;
 
-import com.devcomanda.easymeetup.entity.Meetup;
+import com.devcomanda.easymeetup.model.entity.Meetup;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
