+++ src/main/java/com/devcomanda/easymeetup/service/MeetupServiceImpl.java
@@ -1,7 +1,7 @@
 package com.devcomanda.easymeetup.service;
 
-import com.devcomanda.easymeetup.entity.Meetup;
-import com.devcomanda.easymeetup.entity.exceptions.MeetupNotFoundException;
+import com.devcomanda.easymeetup.model.entity.Meetup;
+import com.devcomanda.easymeetup.model.entity.exceptions.MeetupNotFoundException;
 import com.devcomanda.easymeetup.repository.MeetupRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
