+++ src/test/java/com/devcomanda/easymeetup/controller/MeetupControllerTest.java
@@ -1,6 +1,6 @@
 package com.devcomanda.easymeetup.controller;
 
-import com.devcomanda.easymeetup.entity.Meetup;
+import com.devcomanda.easymeetup.model.entity.Meetup;
 import com.devcomanda.easymeetup.factories.MeetupsFactory;
 import com.devcomanda.easymeetup.service.MeetupService;
 import com.devcomanda.easymeetup.service.security.jwt.TokenProvider;
